+++ scripts/src/Setup/jsc-check.ts
@@ -1,5 +1,6 @@
 import JFS from 'Library/JFS';
 import Job from 'Library/Job';
+import Script from 'Library/JFS/Core/Script';
 
 const job = new Job<string>({
   name: 'jsc-api-check',
@@ -23,4 +24,9 @@ const job = new Job<string>({
   }
 });
 
-job.run();
+export default new Script({
+  path: __filename,
+  name: 'jsc-check',
+  run: () => job.run(),
+  scope: ['all']
+});